requirements.txt
asgiref==3.7.2
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0